Delete stand-alone implement
